Restructuring api folder to be more explicit and clear of content.
